requirements.txt
azure-functions
azure-identity
azure-ai-projects
requests
openpyxl
